pyproject.toml
[project]
name = "k8s-observability-mcp"
version = "0.1.0"
description = "MCP server for Kubernetes observability with Prometheus, Jaeger, Neo4j, logs, and controlled shell tools."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Emre Unlu" }
]
dependencies = [
  "fastmcp>=2.0.0",
  "kubernetes>=30.1.0",
  "requests>=2.32.0",
  "neo4j>=5.28.0",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.0",
  "mypy>=1.10.0",
]

[project.scripts]
k8s-observability-mcp = "main:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main"]

[tool.setuptools.packages.find]
include = ["app*", "clients*", "services*", "tools*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true